Get ~/.composer under version control
PHP may be a horrible tool with disastrous design flaws that make me
weep on an hourly basis, but at the end of the day, I've gotten paid to
work in it for much of my career.

Thus, another step towards making PHP no more painful than it has to be.