Leave emails from the Gnu Emacs Info mailing list in my inbox
I completely missed the release of Emacs 25.2.1 due to that decision.

In retrospect, it was stupid - I *don't* want to see all the noise of
discussion lists I subscribe to in my baseline inbox, but I *do* want to
be notified whenever something significant happens in a project I use.