Add bin/kick-emacs
It just sends SIGUSR2 to running Emacs.app instances, in hopes of
stopping some of the more intense hangs I've experienced.

We'll see if it works.

I only added the command because I keep forgetting the precise
invocation of this command, but am relatively likely to remember I have
a kick-emacs command for such situations.