Add git-rename-branch script
I had need to do it, including a remote branch.

That was such an annoyingly kludgey operation I decided to automate it,
so I don't fall down the path of writing my own VCS out of sheer nerd
rage.

If I go down that path I want it to be a well-reasoned decision, not a
manifestation of my character flaws. :P

(https://github.con/NateEag/next-vcs)