POC a more abstract data structure for Hammerspoon keybindings
My set of mnemonics is getting large enough it's hard to remember what's
been assigned and what has not.

It would be easier to find unassigned keybindings if all keybindings
lived in one data structure.

This commit shows an approach that would make that viable.

"In the end, callbacks were all he needed."