Add handling of laptop display + dual externals to Hammerspoon config
Historically this was rare for me, but in the era of the COVID-19
pandemic, I have to do a lot more video calls and screenshares.

Popping the laptop open can be handy for those, as it's the only webcam
I've got, and can also be a decent place to watch someone's screen while
I have everything else on the other two monitors.